Add frontend and backend for comments
Created a comments class, and added a list of comments to post class. In
post model, added a method to add comment on the post. This method will
not add comment if content is empty. In post view class, added a section
at bottom that will show comments on the given post. At the top there is
a text field to type comment content, and a button to add it. Comments,
like posts, supports markdown.