Add Search user screen
Added a Search User screen that contains a text field, the screen also
contains a button to search. Added a method to UserDataModel class which
checks if the user exists and return bool. If the user exists, it will
redirect you to his profile page, otherwise it will display a message
saying that the user does not exists.